--- app/views/recipes/_list.html.erb
@@ -1,14 +1,0 @@
-<div data-search-recipes-target="list">
-  <p class="h3">
-    <% if params[:query].blank? %>
-      <%= pluralize recipes.count, 'recipes' %> available
-    <% else %>
-      <%= pluralize recipes.count, 'result' %> for <code><%= params[:query] %></code>
-    <% end %>
-  </p>
-  <div class="recipes">
-    <% recipes.each do |recipe| %>
-      <%= render 'recipe_infos', recipe: recipe %>
-    <% end %>
-  </div>
-</div>
