+++ app/views/recipes/_recipe_infos.html.erb
@@ -1,0 +1,5 @@
+<div class="cards-margim" data-search-target="list">
+  <% recipes.each do |recipe|%>
+    <%= render "recipe", recipe: recipe %>
+  <% end %>
+</div>
